requirements.txt
google-genai>=0.3.0
tabulate>=0.9.0
